remove coupling between StreamApi and Client
