+++ src/index.js
@@ -3,6 +3,7 @@ export { default as Expr } from './Expr'
 export { default as PageHelper } from './PageHelper'
 export * from './query'
 export { default as RequestResult } from './RequestResult'
+export { default as StreamAPI } from './stream'
 
 import * as _clientLogger from './clientLogger'
 import * as _errors from './errors'
