+++ src/types/Stream.d.ts
@@ -1,6 +1,17 @@
-import Client from './Client'
-import { ExprArg } from './query'
-import { values } from './values'
+import Expr from './Expr'
+
+type StreamEventFields = ['action', 'document', 'diff', 'prev']
+
+type StreamFn = (
+  expr: Expr,
+  options?: {
+    fields?: StreamEventFields[]
+  }
+) => Subscription
+
+export interface StreamApi extends StreamFn {
+  document: StreamFn
+}
 
 export interface Subscription {
   on: <T extends keyof SubscriptionEventHandlers>(
